Added and validated negative test cases for Product API
- Implemented negative test for invalid endpoint `/product` → expecting 404 response
- Implemented negative test for malformed product ID `/products%7B-9%7D` → expecting 404 HTML error
- Verified proper error handling and response validation (content-type, error message assertions)
- Fixed ClassCastException in positive test by correcting data type handling
- Cleaned up helper methods and improved logging in ProductTests
- Confirmed all positive tests passing successfully except controlled negative validations

+++ src/test/java/paylaods/Payload.java
@@ -1,7 +1,7 @@
 package paylaods;
-import com.github.javafaker.Faker;
 import pojo.Product;
-
+import net.datafaker.Faker;
+import java.math.BigDecimal;
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
 
@@ -18,7 +18,8 @@ public final class Payload {
     public static Product productPayload(){
 
         String title = faker.commerce().productName();
-        double price = faker.number().randomDouble(2, 1, 1000);
+        BigDecimal price = BigDecimal.valueOf(faker.number().randomDouble(2, 1, 1000));
+
         String description = faker.lorem().sentence();
         String image = "https://i.pravatar.cc/300";
